+++ docs/docs/changelog/v0.5.0.md
@@ -116,6 +116,7 @@
 - All images are now converted to .webp for better compression
 
 ### Behind the Scenes
+- The database layer has been added for future recipe scaling. 
 - Black and Flake8 now run as CI/CD checks
 - New debian based docker image
 - Unified Sidebar Components
