+++ tests/utils/helpers.py
@@ -1,0 +1,3 @@
+class MatchAny:
+    def __eq__(self, _: object) -> bool:
+        return True
